Move post processing to postproc
