+++ PyGMO/util/coco/postproc/ppfigparam.py
@@ -18,8 +18,8 @@ import sys
 import matplotlib.pyplot as plt
 import numpy as np
 from pdb import set_trace
-from PyGMO.util._bbob_pproc import toolsstats
-from PyGMO.util._bbob_pproc.ppfig import saveFigure, groupByRange
+from PyGMO.util.coco.postproc import toolsstats
+from PyGMO.util.coco.postproc.ppfig import saveFigure, groupByRange
 
 __all__ = ['beautify', 'plot', 'main']
